Add submission form and download button
There is a distinct lack of form validation, and the UI itself is about
as ugly as you can get. The inputs can mostly fallback to the server
fairly well (minus a malformed date, for example), but time spent on
that would leave no chance of getting this deployed in time.